+++ src/widgets/SearchWidget.php
@@ -50,7 +50,7 @@ class SearchWidget extends Widget
     /**
      * @inheritdoc
      */
-    public static function iconPath()
+    public static function icon()
     {
         return Craft::getAlias('@dukt/twitter/icons/twitter.svg');
     }